Post detection modal loaded to Redux and loaded from there in App3 file
